moved tests to a dedicated crate (marked publish = false)
