+++ Cargo.toml
@@ -3,4 +3,5 @@ members = [
     "http-cache",
     "http-cache-reqwest",
     "http-cache-surf",
+    "http-cache-tests",
 ]
