+++ http-cache-reqwest/src/lib.rs
@@ -173,83 +173,3 @@ impl<T: CacheManager + 'static + Send + Sync> reqwest_middleware::Middleware
         Ok(converted)
     }
 }
-
-#[cfg(feature = "manager-cacache")]
-#[cfg(test)]
-mod tests {
-    use crate::*;
-    use mockito::mock;
-    use reqwest::{Client, Url};
-    use reqwest_middleware::ClientBuilder;
-
-    #[tokio::test]
-    async fn default_mode() -> anyhow::Result<()> {
-        let m = mock("GET", "/")
-            .with_status(200)
-            .with_header("cache-control", "max-age=86400, public")
-            .with_body("test")
-            .create();
-        let url = format!("{}/", &mockito::server_url());
-        let manager = CACacheManager::default();
-        let path = manager.path.clone();
-        let key = format!("GET:{}", &url);
-
-        // Make sure the record doesn't already exist
-        manager.delete("GET", &Url::parse(&url)?).await?;
-
-        // Construct reqwest client with cache defaults
-        let client = ClientBuilder::new(Client::new())
-            .with(Cache(HttpCache {
-                mode: CacheMode::Default,
-                manager: CACacheManager::default(),
-                options: None,
-            }))
-            .build();
-
-        // Cold pass to load cache
-        client.get(url).send().await?;
-        m.assert();
-
-        // Try to load cached object
-        let data = cacache::read(&path, &key).await;
-        assert!(data.is_ok());
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn default_mode_with_options() -> anyhow::Result<()> {
-        let m = mock("GET", "/")
-            .with_status(200)
-            .with_header("cache-control", "max-age=86400, private")
-            .with_body("test")
-            .create();
-        let url = format!("{}/", &mockito::server_url());
-        let manager = CACacheManager::default();
-        let path = manager.path.clone();
-        let key = format!("GET:{}", &url);
-
-        // Make sure the record doesn't already exist
-        manager.delete("GET", &Url::parse(&url)?).await?;
-
-        // Construct reqwest client with cache options override
-        let client = ClientBuilder::new(Client::new())
-            .with(Cache(HttpCache {
-                mode: CacheMode::Default,
-                manager: CACacheManager::default(),
-                options: Some(CacheOptions {
-                    shared: false,
-                    ..Default::default()
-                }),
-            }))
-            .build();
-
-        // Cold pass to load cache
-        client.get(url).send().await?;
-        m.assert();
-
-        // Try to load cached object
-        let data = cacache::read(&path, &key).await;
-        assert!(data.is_ok());
-        Ok(())
-    }
-}
